add_snapshot_options
