+++ cmd/common/common.cpp
@@ -102,15 +102,6 @@ void add_option_wait_mode(CLI::App& cli, silkworm::rpc::WaitMode& wait_mode) {
         ->default_val(silkworm::rpc::WaitMode::blocking);
 }
 
-void add_snapshot_options(CLI::App& cli, SnapshotSettings& snapshot_settings) {
-    cli.add_flag("--snapshots.enabled", snapshot_settings.enabled,
-                 "Flag indicating if usage of snapshots should be enabled or disable");
-    cli.add_flag("--snapshots.no_downloader", snapshot_settings.no_downloader,
-                 "If set, the snapshot downloader is disabled and just already present local snapshots are used");
-
-    // TODO(canepat) add options for the other snapshot settings and for all bittorrent settings
-}
-
 std::string get_node_name_from_build_info(const buildinfo* build_info) {
     std::string node_name{"silkworm/"};
     node_name.append(build_info->git_branch);
